requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
minio==7.1.14
apache-airflow-providers-docker==3.7.1
scikit-learn==1.6.0